dialog_library - made DialogGraphics abstract and extended off it with StoryDialogGraphics (like with the controller). Now adding the DialogGraphics on-click listeners programatically instead of manually (way better with the custom game-specific DialogGraphics-extended class)
