+++ Assets/Scripts/DialogSystem/DialogController.cs
@@ -15,7 +15,7 @@ public class DialogController : MonoBehaviour {
 
     //A handle to the graphics script on the panel.
     //Should be set in the inspector
-    public DialogGraphics graphics;
+    public StoryDialogGraphics graphics;
     public StoryConditionManager conditionManager;
 
     //TODO should make a model for an "Active Conversation" with this info.
@@ -27,6 +27,7 @@ public class DialogController : MonoBehaviour {
     public static event Action<Conversation, bool> StopConversationEvent;
 
     public void Start() {
+        //Listen for when the graphics component tells us that a button has been clicked.
         graphics.SetNextStateEvent += ChooseOption;
     }
 
@@ -118,14 +119,14 @@ public class DialogController : MonoBehaviour {
     //The player has chosen a speech option - update the conversation to reflect this choice
     //If the player chose a special coloured trigger, then show the dramatic dialog before advancing the conversation
     private void ChooseOption(Conversation_Option _option) {
-        Debug.Log("ChooseOption()");
+        Debug.Log("ChooseOption() - headed to state "+_option.target);
 
         conditionManager.HandleTriggers(_option.triggers);
         Conversation_Trigger specialTrigger = conditionManager.getSpecialTrigger(_option.triggers);
 
         //Get the graphics to update - this is a special state now
         if (specialTrigger != null) {
-            graphics.handleTrigger(specialTrigger);
+            graphics.HandleTrigger(specialTrigger);
         }
 
         //Update the active state+statement index, then try to get the first statement from that state
