+++ Assets/Scripts/Graphics/DialogGraphics.cs
@@ -3,32 +3,15 @@ using System.Collections.Generic;
 using System;
 using UnityEngine;
 using UnityEngine.UI;
+using UnityEngine.Events;
 using TMPro;
 
-public class DialogGraphics : MonoBehaviour {
-    private static string PLAYER_NAME = "Pal";
-
-    private static string DEFAULT_MOOD = "neutral";
+public abstract class DialogGraphics : MonoBehaviour {
+    private const string PLAYER_NAME = "Pal";
+    private const string DEFAULT_MOOD = "neutral";
 
     private static Color FULL_VISIBILITY = new Color(1f, 1f, 1f, 1f);
     private static Color HALF_VISIBILITY = new Color(1f, 1f, 1f, 0.5f);
-
-    //Apparently we have to set Colors with normalized (0-1) values if it's editing an Image from a script
-    //Problem: https://issuetracker.unity3d.com/issues/image-color-cannot-be-changed-via-script-when-image-type-is-set-to-simple
-    //Solution: https://forum.unity.com/threads/ui-image-color-not-updating-correctly.674827/
-    private static Dictionary<ValueTuple<String, String>,Color> TRIGGER_COLOURS =
-        new System.Collections.Generic.Dictionary<ValueTuple<String, String>,Color>{
-        {("red", "good"),       new Color(248f/255f,    25f/255f,   0f,         1f)},
-        {("red", "bad"),        new Color(199f/255f,    120f/255f,  120f/255f,  1f)},
-        {("blue", "good"),      new Color(11f/255f,     3f/255f,    252f/255f,  1f)},
-        {("blue", "bad"),       new Color(148f/255f,    183f/255f,  194f/255f,  1f)},
-        {("yellow", "good"),    new Color(255f/255f,    255f/255f,  51f/255f,   1f)},
-        {("yellow", "bad"),     new Color(219f/255f,    224f/255f,  182f/255f,  1f)},
-        {("purple", "good"),    new Color(177f/255f,    0f,         231f/255f,  1f)},
-        {("purple", "bad"),     new Color(99f/255f,     33f/255f,   133f/255f,  1f)},
-        {("green", "good"),     new Color(0f,           232f/255f,  31f/255f,   1f)},
-        {("green", "bad"),      new Color(209f/255f,    237f/255f,  119f/255f,  1f)}
-    };
 
     public Text playerNameplate;
     public Text npcNameplate;
@@ -48,7 +31,7 @@ public class DialogGraphics : MonoBehaviour {
     public event Action<Conversation_Option> SetNextStateEvent;
     
 
-    //Handle activation/deactivation here so that the logic can be separate fromm DialogController
+    //Handle activation/deactivation here so that the logic can be separate from DialogController
     public void activate() { childPanel.gameObject.SetActive(true); }
     public void deactivate() { childPanel.gameObject.SetActive(false); }
 
@@ -83,15 +66,37 @@ public class DialogGraphics : MonoBehaviour {
             Debug.LogError("No conversation options available");
         } else {
             Debug.Log("Player has options to choose from: " + _options);
+            foreach (Conversation_Option option in _options) {
+                Debug.Log(option.optionText);
+            }
 
-            //Fill up the option buttons with the options we've been given
+            //Fill up the option buttons with the Conversation_Options we've been given
             for (int i = 0; i < _options.Count; i++) {
-                //Add a reference on the button to the text (needed for onClick functionality)
+                //Tell the button what Conversation_Option object it's holding
                 playerTextButtonBoxes[i].GetComponent<OptionHolder>().option = _options[i];
+
+                //Display the text contained in that Conversation_Option
                 playerTextButtonBoxes[i].GetComponentInChildren<Text>().text = _options[i].optionText;
+
+                /*
+                Add the event listener to handle the player clicking this button
+                See https://stackoverflow.com/questions/271440/captured-variable-in-a-loop-in-c-sharp for this closure index magic
+                */
+                int closureIndex = i+1;
+                playerTextButtonBoxes[i].GetComponent<Button>().onClick.AddListener(() => onOptionButtonClicked(closureIndex));
                 playerTextButtonBoxes[i].gameObject.SetActive(true);
             }
         }
+    }
+
+    private void onOptionButtonClicked(int _index) {
+        Debug.Log("Clicked index "+_index);
+        SetNextStateEvent?.Invoke(getOptionFromButton(_index));
+    }
+
+    //A small helper function to get a Conversation_Option object from one of the buttons we've setup
+    private Conversation_Option getOptionFromButton(int _buttonIndex) {
+        return playerTextButtonBoxes[_buttonIndex].GetComponent<OptionHolder>().option;
     }
 
     //Player has chosen a dialog option, so show them speaking it
@@ -122,7 +127,7 @@ public class DialogGraphics : MonoBehaviour {
         speakStatement(_npcStatement);
     }
 
-    //Speakerless statement - for use in special colour trigger states
+    //Speakerless statement - fades out both the player and the NPC
     public void speakerlessStatement(Conversation_Statement _statement) {
         //Update the character image fading
         activePlayerImage.color = HALF_VISIBILITY;
@@ -131,35 +136,11 @@ public class DialogGraphics : MonoBehaviour {
         speechText.text = _statement.text;
     }
 
-    public void handleTrigger(Conversation_Trigger _trigger) {
-        resetTextElements();
-
-        //Update the background's colour to reflect the current colour trigger
-        //TODO future - have something more dramatic. Fade-in, not just the default background image, etc
-        childPanel.GetComponent<Image>().color = getColorByTrigger(_trigger);
-        Debug.Log(childPanel.GetComponent<Image>().color);
-    }
-
     //Reset the colour back to normal, in case there was a colour trigger before
     //Used at the beginning of a new state, rather than a new statement, but
     //  DialogGraphics doesn't need to know about states.
     public void resetBackgroundColour() {
         childPanel.GetComponent<Image>().color = FULL_VISIBILITY;
-    }
-
-    //Used as the OnClick function by the buttons created when a player has a choice
-    public void onOptionButtonClicked(int _index) {
-        Debug.Log("Now we should go to state "+playerTextButtonBoxes[_index].GetComponent<OptionHolder>().option.target);
-        SetNextStateEvent?.Invoke(getOptionFromButton(_index));
-    }
-
-    //A small helper function to get a Conversation_Option object from one of the buttons we've setup
-    private Conversation_Option getOptionFromButton(int _buttonIndex) {
-        return playerTextButtonBoxes[_buttonIndex].GetComponent<OptionHolder>().option;
-    }
-
-    private static Color getColorByTrigger(Conversation_Trigger _trigger) {
-        return TRIGGER_COLOURS[(_trigger.colour, _trigger.quality)];
     }
 
     //Update the main text body using a new statement
@@ -175,7 +156,7 @@ public class DialogGraphics : MonoBehaviour {
         speechText.text = _statement.text;
     }
 
-    private void resetTextElements() {
+    protected void resetTextElements() {
         //Disable all the character art + nameplates
         activePlayerImage.enabled = false;
         activeNPCImage.enabled = false;
@@ -194,6 +175,8 @@ public class DialogGraphics : MonoBehaviour {
 
         speechText.text = "";
     }
+
+    public abstract void HandleTrigger(Conversation_Trigger _trigger);
 
     /*
     //How many seconds to wait for the player to read each character of text
