+++ Assets/Scripts/StoryDialogGraphics.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e565c24386141f148aa77416167cc5fd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
